NexGen_Business_Dev/requirements.txt

Flask
python-dotenv